refactor: be more explicit about numeric precision
NetCDF SAM dataset stores the data as 32-bit floats, hence it may be easy to
accidently end up with mixed-precision computations. To avoid that we
need to be more explicit about the dtype of the arrays.

At the moment all computation will be done in 64-bit floats, but this
may be revisisted at a later point.