Strengthen `compute_mixing_length` (#13)
* doc: add docstring for mixing_length module

Does not provide much insight but will allow to introduce and document
module attributes down the line.

* test: test compute_mixing_length against reference values

Reference inputs and outputs of the mixing length computation have been
logged in CLUBB standalone run. They need to be provided to the tests as
an external dataset.

* refactor: use CLUBB constants in the Python calculation

* feat: support multiple saturation formulae

Covers all options avaliable in CLUBB, but only selected are implemented
at the moment.

* fix: incorrect range in Lscale_down computation

The final value of a Python range is exclusive, but to match Fortran
index `1` also needs to be executed.

* fix: uncomment a line to match Fortran reference

Equivalent operation is done in the reference function.