Refactor pre-processing code (#8)
* feat: function to load NetCDF file to Xarray dataset

* feat: add class to provide interface for working with SAM data

* feat: add class to abstract construction of CLUBB grids

One will be expected to build a dataclass using one of the class
constructor methods. The class is immutable hence we can assure that the
grids are consistent and stay like that after construction.

In future we will add extra constructor methods to allow building grids from different inputs.

* feat: project SAM variables to CLUBB like grids

This should enable CLUBB grids being specified independently of the SAM
grid.

The projection assumes that SAM and CLUBB values are piecewise-constant.

---------

Co-authored-by: Jack Atkinson <109271713+jatkinson1000@users.noreply.github.com>